refactor and clean up tests
