+++ SharedLibs/tests/my_lib_1_empty/lib.cpp
