+++ SharedLibs/tests/my_lib_2/lib.hpp
