+++ SharedLibs/tests/my_lib_3_variables_and_init/lib.cpp
@@ -2,7 +2,7 @@
 // Created by Utilisateur on 16/04/2022.
 //
 
-#include "my_lib_3_variables_and_init.hpp"
+#include "lib.hpp"
 
 #include <fstream>
