+++ SharedLibs/tests/my_lib_3_variables_and_init/lib.hpp
@@ -2,8 +2,8 @@
 // Created by Utilisateur on 16/04/2022.
 //
 
-#ifndef MFRANCESCHI_CPPLIBRARIES_MY_LIB_3_VARIABLES_AND_INIT_HPP
-#define MFRANCESCHI_CPPLIBRARIES_MY_LIB_3_VARIABLES_AND_INIT_HPP
+#ifndef MFRANCESCHI_CPPLIBRARIES_LIB_HPP
+#define MFRANCESCHI_CPPLIBRARIES_LIB_HPP
 
 #include "MF/SharedLibsMacros.hpp"
 
@@ -18,4 +18,4 @@ MF_SHARED_LIB_DIRECTIVE extern int thevariable;
 MF_SHARED_LIB_DIRECTIVE extern double theVariable;
 };
 
-#endif // MFRANCESCHI_CPPLIBRARIES_MY_LIB_3_VARIABLES_AND_INIT_HPP
+#endif // MFRANCESCHI_CPPLIBRARIES_LIB_HPP
