+++ SharedLibs/tests/my_lib_5_load_throws/lib.cpp
